Add HandEvaluator class and unit test; change code bewtween HandEvaluator and HandRanking
